Move Stock::Package to OrderManagement::Stock::Package
